currency_converter_project/requirements.txt
Flask==2.3.3
requests==2.31.0
paypalrestsdk==1.13.3
python-dotenv==1.0.1
